feat(auth): configure environment for OAuth secrets
- Add comprehensive .env.example with all OAuth provider variables
- Update env.ts Zod schema to validate OAuth environment variables
- Add OAuth provider setup documentation to README.md
- Support partial OAuth configurations with optional validation
