+++ apps/api/src/core/logger.ts
@@ -1,15 +1,6 @@
 import pino from 'pino';
 import { env, NodeEnvs } from '../env.js';
-
-export const LogLevels = {
-  FATAL: 'fatal',
-  ERROR: 'error',
-  WARN: 'warn',
-  INFO: 'info',
-  DEBUG: 'debug',
-  TRACE: 'trace',
-  SILENT: 'silent',
-} as const;
+import { LogLevels } from './types.js';
 
 // Define sensitive fields that should be redacted in logs
 const REDACT_FIELDS = [
