+++ apps/api/src/core/types.ts
@@ -1,0 +1,9 @@
+export const LogLevels = {
+  FATAL: 'fatal',
+  ERROR: 'error',
+  WARN: 'warn',
+  INFO: 'info',
+  DEBUG: 'debug',
+  TRACE: 'trace',
+  SILENT: 'silent',
+} as const;
